fix(xdesign-story): 파일구조 변경
