+++ services/xdesign-story/src/foundations/Color/Color.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, ColorPalette, ColorItem } from '@storybook/blocks';
-import { colors } from '..';
+import { colors } from './';
 
 <Meta title="FOUNDATIONS/Colors" />
