+++ services/xdesign-story/src/foundations/Color/Color.ts
@@ -1,0 +1,289 @@
+import { ColorTheme } from './Color.types';
+
+export const auto: Readonly<ColorTheme> = {
+    scheme: {
+        white: 'var(--white)',
+        black: 'var(--black)',
+        /* primary */
+        primary0: 'var(--primary0)',
+        primary10: 'var(--primary10)',
+        primary20: 'var(--primary20)',
+        primary30: 'var(--primary30)',
+        primary40: 'var(--primary40)',
+        primary50: 'var(--primary50)',
+        primary60: 'var(--primary60)',
+        primary70: 'var(--primary70)',
+        primary80: 'var(--primary80)',
+        primary90: 'var(--primary90)',
+        primary95: 'var(--primary95)',
+        primary99: 'var(--primary99)',
+        primary100: 'var(--primary100)',
+
+        /* secondary */
+        secondary0: 'var(--secondary0)',
+        secondary10: 'var(--secondary10)',
+        secondary20: 'var(--secondary20)',
+        secondary30: 'var(--secondary30)',
+        secondary40: 'var(--secondary40)',
+        secondary50: 'var(--secondary50)',
+        secondary60: 'var(--secondary60)',
+        secondary70: 'var(--secondary70)',
+        secondary80: 'var(--secondary80)',
+        secondary90: 'var(--secondary90)',
+        secondary95: 'var(--secondary95)',
+        secondary99: 'var(--secondary99)',
+        secondary100: 'var(--secondary100)',
+
+        /* tertiary */
+        tertiary0: 'var(--tertiary0)',
+        tertiary10: 'var(--tertiary10)',
+        tertiary20: 'var(--tertiary20)',
+        tertiary30: 'var(--tertiary30)',
+        tertiary40: 'var(--tertiary40)',
+        tertiary50: 'var(--tertiary50)',
+        tertiary60: 'var(--tertiary60)',
+        tertiary70: 'var(--tertiary70)',
+        tertiary80: 'var(--tertiary80)',
+        tertiary90: 'var(--tertiary90)',
+        tertiary95: 'var(--tertiary95)',
+        tertiary99: 'var(--tertiary99)',
+        tertiary100: 'var(--tertiary100)',
+
+        /* error */
+        error0: 'var(--error0)',
+        error10: 'var(--error10)',
+        error20: 'var(--error20)',
+        error30: 'var(--error30)',
+        error40: 'var(--error40)',
+        error50: 'var(--error50)',
+        error60: 'var(--error60)',
+        error70: 'var(--error70)',
+        error80: 'var(--error80)',
+        error90: 'var(--error90)',
+        error95: 'var(--error95)',
+        error99: 'var(--error99)',
+        error100: 'var(--error100)',
+
+        /* neutral */
+        neutral0: 'var(--neutral0)',
+        neutral10: 'var(--neutral10)',
+        neutral20: 'var(--neutral20)',
+        neutral30: 'var(--neutral30)',
+        neutral40: 'var(--neutral40)',
+        neutral50: 'var(--neutral50)',
+        neutral60: 'var(--neutral60)',
+        neutral70: 'var(--neutral70)',
+        neutral80: 'var(--neutral80)',
+        neutral90: 'var(--neutral90)',
+        neutral95: 'var(--neutral95)',
+        neutral99: 'var(--neutral99)',
+        neutral100: 'var(--neutral100)',
+
+        /* neutral_variant */
+        neutral_variant0: 'var(--neutral_variant0)',
+        neutral_variant10: 'var(--neutral_variant10)',
+        neutral_variant20: 'var(--neutral_variant20)',
+        neutral_variant30: 'var(--neutral_variant30)',
+        neutral_variant40: 'var(--neutral_variant40)',
+        neutral_variant50: 'var(--neutral_variant50)',
+        neutral_variant60: 'var(--neutral_variant60)',
+        neutral_variant70: 'var(--neutral_variant70)',
+        neutral_variant80: 'var(--neutral_variant80)',
+        neutral_variant90: 'var(--neutral_variant90)',
+        neutral_variant95: 'var(--neutral_variant95)',
+        neutral_variant99: 'var(--neutral_variant99)',
+        neutral_variant100: 'var(--neutral_variant100)',
+    },
+};
+
+export const light: Readonly<ColorTheme> = {
+    scheme: {
+        white: '#000000',
+        black: '#FFFFFF',
+        /* primary */
+        primary0: '#000000',
+        primary10: '#1D0D33',
+        primary20: '#35185C',
+        primary30: '#4D2485',
+        primary40: '#6532AD',
+        primary50: '#7E40D5',
+        primary60: '#9650FA',
+        primary70: '#AA71FB',
+        primary80: '#BE93FC',
+        primary90: '#D2B4FD',
+        primary95: '#E6D5FE',
+        primary99: '#FAF6FF',
+        primary100: '#FFFFFF',
+
+        /* secondary */
+        secondary0: '#000000',
+        secondary10: '#130C33',
+        secondary20: '#22175B',
+        secondary30: '#332382',
+        secondary40: '#4430A8',
+        secondary50: '#563ECC',
+        secondary60: '#684DF0',
+        secondary70: '#856FF3',
+        secondary80: '#A191F6',
+        secondary90: '#BEB3F9',
+        secondary95: '#DBD4FB',
+        secondary99: '#F8F6FE',
+        secondary100: '#FFFFFF',
+
+        /* tertiary */
+        tertiary0: '#000000',
+        tertiary10: '#121531',
+        tertiary20: '#212658',
+        tertiary30: '#30387D',
+        tertiary40: '#414BA2',
+        tertiary50: '#535EC6',
+        tertiary60: '#6572E9',
+        tertiary70: '#828DED',
+        tertiary80: '#A0A8F1',
+        tertiary90: '#BDC2F6',
+        tertiary95: '#DADDFA',
+        tertiary99: '#F7F8FE',
+        tertiary100: '#FFFFFF',
+
+        /* error */
+        error0: '#000000',
+        error10: '#410E0B',
+        error20: '#601410',
+        error30: '#8C1D18',
+        error40: '#B3261E',
+        error50: '#DC362E',
+        error60: '#E46962',
+        error70: '#EC928E',
+        error80: '#F2B8B5',
+        error90: '#F9DEDC',
+        error95: '#FCEEEE',
+        error99: '#FFFBF9',
+        error100: '#FFFFFF',
+
+        /* neutral */
+        neutral0: '#000000',
+        neutral10: '#1C1B1F',
+        neutral20: '#313033',
+        neutral30: '#484649',
+        neutral40: '#605D62',
+        neutral50: '#787579',
+        neutral60: '#939094',
+        neutral70: '#AEAAAE',
+        neutral80: '#C9C5CA',
+        neutral90: '#E6E1E5',
+        neutral95: '#F4EFF4',
+        neutral99: '#FFFBFE',
+        neutral100: '#FFFFFF',
+
+        /* neutral_variant */
+        neutral_variant0: '#000000',
+        neutral_variant10: '#1D1A22',
+        neutral_variant20: '#322F37',
+        neutral_variant30: '#49454F',
+        neutral_variant40: '#605D66',
+        neutral_variant50: '#79747E',
+        neutral_variant60: '#938F99',
+        neutral_variant70: '#AEA9B4',
+        neutral_variant80: '#CAC4D0',
+        neutral_variant90: '#E7E0EC',
+        neutral_variant95: '#F5EEFA',
+        neutral_variant99: '#FFFBFE',
+        neutral_variant100: '#FFFFFF',
+    },
+};
+
+export const dark: Readonly<ColorTheme> = {
+    scheme: {
+        white: '#000000',
+        black: '#FFFFFF',
+        /* primary */
+        primary0: '#000000',
+        primary10: '#1D0D33',
+        primary20: '#35185C',
+        primary30: '#4D2485',
+        primary40: '#6532AD',
+        primary50: '#7E40D5',
+        primary60: '#9650FA',
+        primary70: '#AA71FB',
+        primary80: '#BE93FC',
+        primary90: '#D2B4FD',
+        primary95: '#E6D5FE',
+        primary99: '#FAF6FF',
+        primary100: '#FFFFFF',
+
+        /* secondary */
+        secondary0: '#000000',
+        secondary10: '#130C33',
+        secondary20: '#22175B',
+        secondary30: '#332382',
+        secondary40: '#4430A8',
+        secondary50: '#563ECC',
+        secondary60: '#684DF0',
+        secondary70: '#856FF3',
+        secondary80: '#A191F6',
+        secondary90: '#BEB3F9',
+        secondary95: '#DBD4FB',
+        secondary99: '#F8F6FE',
+        secondary100: '#FFFFFF',
+
+        /* tertiary */
+        tertiary0: '#000000',
+        tertiary10: '#121531',
+        tertiary20: '#212658',
+        tertiary30: '#30387D',
+        tertiary40: '#414BA2',
+        tertiary50: '#535EC6',
+        tertiary60: '#6572E9',
+        tertiary70: '#828DED',
+        tertiary80: '#A0A8F1',
+        tertiary90: '#BDC2F6',
+        tertiary95: '#DADDFA',
+        tertiary99: '#F7F8FE',
+        tertiary100: '#FFFFFF',
+
+        /* error */
+        error0: '#000000',
+        error10: '#410E0B',
+        error20: '#601410',
+        error30: '#8C1D18',
+        error40: '#B3261E',
+        error50: '#DC362E',
+        error60: '#E46962',
+        error70: '#EC928E',
+        error80: '#F2B8B5',
+        error90: '#F9DEDC',
+        error95: '#FCEEEE',
+        error99: '#FFFBF9',
+        error100: '#FFFFFF',
+
+        /* neutral */
+        neutral0: '#000000',
+        neutral10: '#1D1A22',
+        neutral20: '#313033',
+        neutral30: '#484649',
+        neutral40: '#605D62',
+        neutral50: '#787579',
+        neutral60: '#939094',
+        neutral70: '#AEAAAE',
+        neutral80: '#C9C5CA',
+        neutral90: '#E6E1E5',
+        neutral95: '#F4EFF4',
+        neutral99: '#FFFBFE',
+        neutral100: '#FFFFFF',
+
+        /* neutral_variant */
+        neutral_variant0: '#000000',
+        neutral_variant10: '#1D1A22',
+        neutral_variant20: '#322F37',
+        neutral_variant30: '#49454F',
+        neutral_variant40: '#605D66',
+        neutral_variant50: '#79747E',
+        neutral_variant60: '#938F99',
+        neutral_variant70: '#AEA9B4',
+        neutral_variant80: '#CAC4D0',
+        neutral_variant90: '#E7E0EC',
+        neutral_variant95: '#F5EEFA',
+        neutral_variant99: '#FFFBFE',
+        neutral_variant100: '#FFFFFF',
+    },
+};
