--- services/xdesign-story/src/foundations/index.ts
@@ -1,3 +1,0 @@
-export * as colors from './Color';
-
-export type { ColorTheme, ColorScheme, KnownColorGroup, ColorToken } from './Color/types';
